requirements.txt
asgiref==3.10.0
channels>=4.0
channels-redis>=4.0
daphne>=4.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
Pillow==11.0.0
sqlparse==0.5.3
